pyproject.toml
[tool.poetry]
name = "credit-scoring-system"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ray = {extras = ["serve"], version = "^2.6.3"}
fastapi = "^0.103.1"
pydantic = "1.10"
poetry-dotenv-plugin = "^0.2.0"


[tool.poetry.group.testing.dependencies]
flake8 = "^6.1.0"
black = "^23.9.1"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
